Make E*Trade ESPP purchases deliveries
The type of E*Trade (aka Morgan Stanley at Work) accounts that are used
of Employee Stock Purchase Programs (ESPP) which the supported "purchase"
document is for do not even allow regular purchases done by the account
holder. All these transactions are triggered by the employer of the
account holder and no money is ever taken from the associated deposit
account, it is deducted from the employees pay slip.
Therefore, it makes more sense to treat these as a delivery and spare
the user having to tick the according box on every import.