Handle short company name not split across two lines
The document has 2 columns and if the company name is short enough
to not be word-wrapped, it uses up only one line in the first column
instead of two as for the existing examlpe.
This change adds anoher example where this is the case and adapts
the matching of the lines accordingly to handle both cases.