Handle Employee Stock Option Exercise confirmation documents
The "Order Summary" part of the document is 2 columns, the "Exercise Details"
part is single column an duplicates quite some of the info.
We use "Shares Issued" instead of "Shares Exercised" to determine the amount
of shares deposited into the account, based on the assumption, that for an
exercise of type "Sell-to-Cover", where part of the shares are immediately sold
by the broker to cover exercise cost, the former will show the number of
remaining shares.